assistant-client/requirements.txt
aiohttp==3.8.4
pvrecorder==1.1.1
python-dotenv==1.0.0
soundfile==0.12.1
pydub==0.25.1
pvporcupine==2.2.0
requests==2.28.2
websockets==11.0.2
PyYAML==6.0
PyAudio==0.2.13

